Move MapAny/Align/Pack widgets to new kas_core::widgets::adapt module
