+++ crates/kas-macros/src/collection.rs
@@ -172,8 +172,8 @@ impl Item {
                 let _ = parenthesized!(inner in input);
                 let hints: Expr = inner.parse()?;
 
-                ty = quote! { ::kas::hidden::Align<#ty> };
-                def = quote! { ::kas::hidden::Align::new(#def, #hints) };
+                ty = quote! { ::kas::widgets::adapt::Align<#ty> };
+                def = quote! { ::kas::widgets::adapt::Align::new(#def, #hints) };
             } else if input.peek(Token![.]) && input.peek2(kw::pack) {
                 let _: Token![.] = input.parse()?;
                 let _: kw::pack = input.parse()?;
@@ -182,8 +182,8 @@ impl Item {
                 let _ = parenthesized!(inner in input);
                 let hints: Expr = inner.parse()?;
 
-                ty = quote! { ::kas::hidden::Pack<#ty> };
-                def = quote! { ::kas::hidden::Pack::new(#def, #hints) };
+                ty = quote! { ::kas::widgets::adapt::Pack<#ty> };
+                def = quote! { ::kas::widgets::adapt::Pack::new(#def, #hints) };
             }
 
             Ok(Item::Label(names.next(), ty, def))
@@ -330,8 +330,12 @@ impl Collection {
             let path = match item {
                 Item::Label(stor, ty, def) => {
                     if let Some(ref data_ty) = data_ty {
-                        stor_ty.append_all(quote! { #stor: ::kas::hidden::MapAny<#data_ty, #ty>, });
-                        stor_def.append_all(quote! { #stor: ::kas::hidden::MapAny::new(#def), });
+                        stor_ty.append_all(
+                            quote! { #stor: ::kas::widgets::adapt::MapAny<#data_ty, #ty>, },
+                        );
+                        stor_def.append_all(
+                            quote! { #stor: ::kas::widgets::adapt::MapAny::new(#def), },
+                        );
                     } else {
                         stor_ty.append_all(quote! { #stor: #ty, });
                         stor_def.append_all(quote! { #stor: #def, });
